Re-organise interfaces
