+++ packages/resources/src/interfaces.ts
@@ -1,61 +1,4 @@
-export type Status = 'failed' | 'loading' | 'completed';
-export type Action = 'create' | 'remove' | 'update' | 'read';
-
-export enum ResourceResponseStatus {
-	failed = 0,
-	success = 1
-}
-
-export interface Resource<S> {
-	getOrRead(): S[];
-}
-
-export interface ManyResourceResponse<S> {
-	data: S[];
-	status: ResourceResponseStatus;
-}
-
-export interface ResourceRead<S> {
-	(): Promise<ManyResourceResponse<S[]>>;
-}
-
-export interface ResourceConfig<S> {
-	idKey?: string;
-	template(resource: Partial<S>): S;
-	read: ResourceRead<S>;
-}
-
-// Rest stuff
-
-export interface RestResourceUrlOptions {
-	origin: string;
-	name: string;
-	id?: string;
-}
-
-export interface RestResourceUrlFunction {
-	(options: RestResourceUrlOptions): string;
-}
-
-export interface RestResourceOperationsConfig {
-	optimistic: boolean;
-	verb: 'GET' | 'POST' | 'PATCH' | 'DELETE' | 'PUT';
-	url: RestResourceUrlFunction;
-}
-
-export interface RestManyResourceOperations {
-	read: false | RestResourceOperationsConfig;
-}
-
-export interface RestResource<S> {
-	origin: string;
-	name: string;
-	idKey?: string;
-	template?(resource: Partial<S>): S;
-	many?: Partial<RestManyResourceOperations>;
-}
-
-// Store State
+import { Status, Action } from './ResourceProvider';
 
 export interface ResourceData<S> {
 	[index: string]: S;
